编写完程序并验证，可以立起来了，但有很多bug，需要大量优化

+++ User/Devices/Dji_motor/Motor_Dji.h
@@ -29,13 +29,13 @@ typedef struct
     int16_t  RawEcd;				//< 原始编码器数据
     int16_t  LastEcd;			  //< 上一时刻编码器返回值
     float    Angle;					//< 解算后的编码器角度
-    int16_t  AngleSpeed;	  //< 解算后的编码器角速度
+    float  AngleSpeed;	  //< 解算后的编码器角速度
     int32_t  RoundCnt;			//< 累计转动圈数
     int32_t  TotalEcd;			//< 编码器累计增量值
     int32_t  TotalAngle;		//< 累计旋转角度
 
     int32_t  Target;				//< 电机的期望参数
-    int32_t  Output;  			//< 电机输出值，通常为电流和电压
+    int16_t  Output;  			//< 电机输出值，通常为电流和电压
     float CanEcd[20] ;
     float CanAngleSpeed[20] ;
     float LvboAngle;
